fixed random crashes in Release build small tweak for firewalled nodes option to autoupdate IP on DHT firewall check all DHT settings should work default DHT port set to 6245 tried adding string pool for storing userinfo updated OpenSSL

+++ client/FinishedManager.cpp
@@ -82,7 +82,7 @@ void FinishedManager::on(QueueManagerListener::Finished, const QueueItem* qi, co
 		FinishedItemPtr item = new FinishedItem(
 			qi->getTarget(), d->getUser(),
 			Util::toString(ClientManager::getInstance()->getHubNames(d->getUser()->getCID())),
-			qi->getSize(), d->getAverageSpeed(), GET_TIME(), qi->getTTH().toBase32());
+			qi->getSize(), static_cast<int64_t>(d->getAverageSpeed()), GET_TIME(), qi->getTTH().toBase32());
 		{
 			Lock l(cs);
 			downloads.push_back(item);
@@ -109,7 +109,7 @@ void FinishedManager::on(UploadManagerListener::Complete, const Upload* u) throw
 		FinishedItemPtr item = new FinishedItem(
 			u->getPath(), u->getUser(),
 			Util::toString(ClientManager::getInstance()->getHubNames(u->getUser()->getCID())),
-			u->getFileSize(), u->getAverageSpeed(), GET_TIME());
+			u->getFileSize(), static_cast<int64_t>(u->getAverageSpeed()), GET_TIME());
 		{
 			Lock l(cs);
 			uploads.push_back(item);
