+++ windows/TransferView.cpp
@@ -914,7 +914,7 @@ void TransferView::on(DownloadManagerListener::Tick, const DownloadList& dl) {
 		ui->setPos(d->getPos());
 		ui->setSize(d->getSize());
 		ui->setTimeLeft(d->getSecondsLeft());
-		ui->setSpeed(d->getAverageSpeed());
+		ui->setSpeed(static_cast<int64_t>(d->getAverageSpeed()));
 
 		tstring pos = Util::formatBytesW(d->getPos());
 		double percent = (double)d->getPos()*100.0/(double)d->getSize();
@@ -1016,7 +1016,7 @@ void TransferView::on(UploadManagerListener::Tick, const UploadList& ul) {
 		ui->setActual(u->getStartPos() + u->getActual());
 		ui->setPos(u->getStartPos() + u->getPos());
 		ui->setTimeLeft(u->getSecondsLeft(true)); // we are interested when whole file is finished and not only one chunk
-		ui->setSpeed(u->getAverageSpeed());
+		ui->setSpeed(static_cast<int64_t>(u->getAverageSpeed()));
 
 		tstring pos = Util::formatBytesW(ui->pos);
 		double percent = (double)ui->pos*100.0/(double)(u->getType() == Transfer::TYPE_TREE ? u->getSize() : u->getFileSize());
@@ -1226,7 +1226,7 @@ void TransferView::on(QueueManagerListener::StatusUpdated, const QueueItem* qi) 
 					chunked = true;
 				}
 		
-				totalSpeed += d->getAverageSpeed();
+				totalSpeed += static_cast<int64_t>(d->getAverageSpeed());
 				ratio += d->getPos() > 0 ? (double)d->getActual() / (double)d->getPos() : 1.00;
 			}
 		}
